fixing mask projection

+++ util/utilIO.cpp
@@ -53,8 +53,8 @@ void ImgIO::getPtsFromMask(const cv::Mat &mask, std::vector<cv::Point2f> &pts_ve
 
 cv::Mat ImgIO::getRtMatrix(const vcg::Shot<float> &shot){
 
-  cv::Mat mat_Rt(3,4, CV_64FC1);
-  mat_Rt = cv::Mat::zeros(3,4, CV_64FC1);
+  cv::Mat mat_Rt(3,4, CV_32FC1);
+  mat_Rt = cv::Mat::zeros(3,4, CV_32FC1);
 
   vcg::Matrix44f mat_rot = shot.Extrinsics.Rot();
   vcg::Point3f mat_tra = shot.Extrinsics.Tra();
@@ -71,13 +71,13 @@ cv::Mat ImgIO::getRtMatrix(const vcg::Shot<float> &shot){
 cv::Mat ImgIO::getIntrMatrix(const vcg::Shot<float> &shot){
 
   cv::Mat intr_mat;
-  intr_mat = cv::Mat::zeros(3,3, CV_64FC1);
+  intr_mat = cv::Mat::zeros(3,3, CV_32FC1);
 
   intr_mat.at<float>(0,0) = shot.Intrinsics.FocalMm;
   intr_mat.at<float>(1,1) = shot.Intrinsics.FocalMm;
   intr_mat.at<float>(0,2) = shot.Intrinsics.CenterPx[0];
   intr_mat.at<float>(1,2) = shot.Intrinsics.CenterPx[1];
-  intr_mat.at<float>(2,2,0) = static_cast<float>(1.0);
+  intr_mat.at<float>(2,2,0) = 1;
 
   std::cout<< intr_mat <<std::endl;
